[hw] Use the built-in IntEnum instead of creating a new class (#5433)
# Description

There is a built-in type IntEnum in Python 3.8 - 3.10, thus no need to
create a new class.

# How Has This Been Tested?
<!-- Please describe the tests that you ran to verify your changes.
Provide instructions so we can reproduce.
-->

- [X] Manuall tested in a local development environment

# Checklist
- [ ] The PR is tagged with proper labels (bug, enhancement, feature,
documentation)
- [X] I have performed a self-review of my own code
- [ ] I have added unit tests that prove my fix is effective or that my
feature works
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation

cc:
<!-- Optionally mention someone to let them know about this pull request
-->
@wangxiaoyou1993
